Bug Fixed

+++ src/it/polimi/tiw/bank/controllers/CreateUser.java
@@ -23,7 +23,7 @@ import java.sql.SQLException;
 @WebServlet("/CreateUser")
 public class CreateUser extends HttpServlet {
 
-    private static final long serialVersionUID = 1L;
+    private static final double serialVersionUID = 1L;
     private Connection connection;
     private TemplateEngine templateEngine;
