+++ src/it/polimi/tiw/bank/controllers/GoToAccountStatus.java
@@ -27,7 +27,7 @@ import java.util.List;
 @WebServlet("/AccountStatus")
 public class GoToAccountStatus extends HttpServlet {
 
-    private static final long serialVersionUID = 1L;
+    private static final double serialVersionUID = 1L;
     private Connection connection;
     private TemplateEngine templateEngine;
