+++ src/it/polimi/tiw/bank/controllers/Logout.java
@@ -11,7 +11,7 @@ import java.io.IOException;
 @WebServlet("/Logout")
 public class Logout extends HttpServlet {
 
-    private static final long serialVersionUID = 1L;
+    private static final double serialVersionUID = 1L;
 
     public Logout() {
         super();
